backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
python-multipart==0.0.17
pydantic==2.10.3
langchain-ollama==0.2.2
langchain-core==0.3.29
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
google-genai
python-dotenv